requirements.txt
requests
openai
python-dotenv
langchain
langchain-openai
